- 레이어 복사를 위한 리소스 복사작업 완료.(+폰트, 폼, 이미지) - 그래픽 상태 처리 관련 일부 오류 수정. - 스트림 처리 관련 오류 수정.

+++ src/Element/Graphic/PPGState.cpp
@@ -166,11 +166,11 @@ string PPGState::MakeCommandString()
     }
     if(_gflag & PPGF_LINECAP) {
         cinfo = &PPCommandList[PPC_LineCap];
-        ostr << _lineCap << " " << cinfo->command << PP_ENDL;
+        ostr << (int)_lineCap << " " << cinfo->command << PP_ENDL;
     }
     if(_gflag & PPGF_LINEJOIN) {
         cinfo = &PPCommandList[PPC_LineJoin];
-        ostr << _lineJoin << " " << cinfo->command << PP_ENDL;
+        ostr << (int)_lineJoin << " " << cinfo->command << PP_ENDL;
     }
     if(_gflag & PPGF_DICTNAME) {
         cinfo = &PPCommandList[PPC_DictName];
